nested key example
This works to significantly slow down discovery of inner code.  This example is
fairly sensible as it shows how to use a relatively unobvious starting key and
then on the inside there can be a number of other environmental keys.  While
this one merely uses the perl version on the inside, one could obviously use
other variables, including a literal keyfile.